pyproject.toml
[tool.poetry]
name = "talent-api-examples"
version = "0.1.0"
description = ""
authors = ["Semyachkin Vitaly <tonatossn@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Authlib = "^0.15"
requests = "^2.24.0"
python-dotenv = "^0.14.0"
pygments = "^2.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
